% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all_units_as_group.R
\name{add_all_units_as_group}
\alias{add_all_units_as_group}
\title{this add all units as a separate group, for grouped data}
\usage{
add_all_units_as_group(df, group_var, new_group_id = "")
}
\arguments{
\item{df}{Data frame}

\item{group_var}{the name of the var that defines groups}

\item{new_group_id}{new id for the group that is composed of all units}
}
\description{
this add all units as a separate group, for grouped data
}
