% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_linebreak.R
\name{add_linebreak}
\alias{add_linebreak}
\title{add linebreaks to strings more than n characters in length}
\usage{
add_linebreak(string, min_length = 10, add_multiple_linebreaks = F)
}
\arguments{
\item{string}{string}

\item{min_length}{minimum length for string to be broken into two lines}

\item{add_multiple_linebreaks}{if set to T, linebreaks will be added to all but the first whitespace or dash}
}
\value{
returns a string
}
\description{
add linebreaks to strings more than n characters in length
}
