% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_linebreak_vector.R
\name{add_linebreak_vector}
\alias{add_linebreak_vector}
\title{add linebreaks to strings more than n characters in length, vectorized}
\usage{
add_linebreak_vector(string, ...)
}
\arguments{
\item{string}{string vector}

\item{...}{arguments passed to add_linebreak}
}
\value{
returns a string vector
}
\description{
add linebreaks to strings more than n characters in length, vectorized
}
