% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_make_runall_file.R
\name{auto_make_runall_file}
\alias{auto_make_runall_file}
\title{This makes a runall file from a directory that has r files}
\usage{
auto_make_runall_file(code_directory, out_file = "", send_to_clipboard = T)
}
\arguments{
\item{code_directory}{Directory that has all the code (needs full name)}

\item{out_file}{outout .R file}

\item{send_to_clipboard}{if T, then output will be sent to clipboard}
}
\value{
This saves an .R file
}
\description{
This makes a runall file from a directory that has r files
}
\examples{
auto_make_runall_file(code_directory = "C:/Users/hanno/Dropbox/Harvard/Projects/Newspapers/02_Code/", 
                     out_file = "02_Code/RUNALL_TEST.R")


}
