% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missings_df.R
\name{check_missings_df}
\alias{check_missings_df}
\title{This gives some info on missingness}
\usage{
check_missings_df(df, varlist, group = "")
}
\arguments{
\item{df}{Data frame}

\item{varlist}{list of variables to check}

\item{group}{Group for grouped output (leave blank if no group)}
}
\description{
This gives some info on missingness
}
