% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlap.R
\name{check_overlap}
\alias{check_overlap}
\title{Checks overlap between two vectors (left and right sided)}
\usage{
check_overlap(vec_x, vec_y, return_miss = F)
}
\arguments{
\item{vec_x}{Vector}

\item{vec_y}{Vector}

\item{return_miss}{If T, return list of vec_x missing from vec_y}
}
\description{
Checks overlap between two vectors (left and right sided)
}
