% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_wd.R
\name{detect_wd}
\alias{detect_wd}
\title{this sets the workig directory automatically for joint projects}
\usage{
detect_wd(
  wd_h = "",
  wd_h2 = "",
  wd_alt = "",
  user_name = "Hanno",
  user_name2 = "hah235"
)
}
\arguments{
\item{wd_h}{my working directory}

\item{wd_h2}{my other directory}

\item{wd_alt}{alternative working directory}

\item{user_name}{name of the OS user as returned by sys.Info()}

\item{user_name2}{name of the OS user as returned by sys.Info()}
}
\value{
this function does not return anything
}
\description{
this sets the workig directory automatically for joint projects
}
\examples{
detect_wd(wd_h = "C:/Users/Hanno/Dropbox/",
          wd_alt = "C:/Users/Hanno/",
          user_name = "Hanno")

}
