% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_missings_df.R
\name{drop_missings_df}
\alias{drop_missings_df}
\title{Returns DF without missings}
\usage{
drop_missings_df(df, varlist)
}
\arguments{
\item{df}{df}

\item{varlist}{list of vars to be considered for dropping missings}
}
\value{
DF with original string, match and distance to matchs
}
\description{
Returns DF without missings
}
