% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feols_residualize.R
\name{feols_residualize}
\alias{feols_residualize}
\title{Return residuals from a fixed effects model}
\usage{
feols_residualize(y, rhs_continuous, rhs_fe, data)
}
\arguments{
\item{y}{character, dependent variable}

\item{rhs_continuous}{character vector, right-hand side continuous variables}

\item{rhs_fe}{character vector, right-hand side fixed effects (basically any categorical variable)}

\item{data}{data.frame, data}
}
\description{
Return residuals from a fixed effects model
}
