% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixest_dict_to_df.R
\name{fixest_dict_to_df}
\alias{fixest_dict_to_df}
\title{Convert fixest dict to DF (requires that fixest dict has been set before)}
\usage{
fixest_dict_to_df()
}
\description{
Convert fixest dict to DF (requires that fixest dict has been set before)
}
