% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_string_join.R
\name{fuzzy_string_join}
\alias{fuzzy_string_join}
\title{Fuzzy string join function}
\usage{
fuzzy_string_join(df_left, df_right, by = "", quantile_dist_not_match = 0.5)
}
\arguments{
\item{df_left}{left df}

\item{df_right}{right df}

\item{by}{name of string to be matched onstringdist}

\item{quantile_dist_not_match}{q of distribution of distance such that greater distances will not be matched}
}
\value{
left df with matched stuff from right df
}
\description{
Fuzzy string join function
}
