% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_to_lon_lat.R
\name{geometry_to_lonlat}
\alias{geometry_to_lonlat}
\title{A function to convert sf objects into data frames including lon/lat columns}
\usage{
geometry_to_lonlat(x)
}
\arguments{
\item{x}{an sf object}
}
\description{
A function to convert sf objects into data frames including lon/lat columns
}
