% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bal.R
\name{get_bal}
\alias{get_bal}
\title{Standardized Covariate Balance}
\usage{
get_bal(treatvar, cov_list, data.df, FE)
}
\arguments{
\item{treatvar}{treatment variable}

\item{cov_list}{list of covariates to assess balance on}

\item{data.df}{data frame storing all variables}

\item{FE}{fixed effects: stratify by this variable when assessing balance}
}
\value{
This function returns the standardized covariate balance
}
\description{
Standardized Covariate Balance
}
