% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coef.R
\name{get_coef}
\alias{get_coef}
\title{this extracts the coefficients from lm}
\usage{
get_coef(model = m1, which.ivar = "Petal.Width")
}
\arguments{
\item{model}{lm object}

\item{which.ivar}{which independet variable should be obtained?}
}
\value{
a vector of estimation results
}
\description{
this extracts the coefficients from lm
}
\examples{
get_coefs(model = lm(Petal.Length ~ Petal.Width, data = iris), which.ivar = "Petal.Width")

}
