% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_check_obs.r
\name{left_join_check_obs}
\alias{left_join_check_obs}
\title{this function checks if the number of observations changes after a left join}
\usage{
left_join_check_obs(x, y, ...)
}
\arguments{
\item{x}{Left df}

\item{y}{right df}

\item{...}{other arguments to pass to left_join}
}
\description{
this function checks if the number of observations changes after a left join
}
