% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary.R
\name{make_binary}
\alias{make_binary}
\title{Convert cont var to binary}
\usage{
make_binary(v, method = "median", q = 0.75)
}
\arguments{
\item{v}{variable (must be continuous)}

\item{method}{mean, median or quantile split}

\item{q}{if method = quantile, then this is the quantile that is used (0-1)}
}
\description{
Convert cont var to binary
}
