% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subgroup_df_list.R
\name{make_hte_dfs}
\alias{make_hte_dfs}
\title{Takes sdata frame and list of vars (categorical), and them makes a list of data frame}
\usage{
make_hte_dfs(htevars, data)
}
\arguments{
\item{htevars}{list of vars that data sets will be based on}

\item{data}{data frame}
}
\value{
Returns  list of data sets and labels
}
\description{
Takes sdata frame and list of vars (categorical), and them makes a list of data frame
}
