% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_plots_to_overleaf.R
\name{move_plots_to_overleaf}
\alias{move_plots_to_overleaf}
\title{find and run MPTO script in a project without having to run script manually}
\usage{
move_plots_to_overleaf(
  code_directory_string = "Code",
  script_string = "Overleaf"
)
}
\arguments{
\item{code_directory_string}{this should be a strong that identifies the code directory}

\item{script_string}{this should be a string that identifies the MPTO script in the code directory}
}
\value{
returns nothing but runs the script
}
\description{
find and run MPTO script in a project without having to run script manually
}
