% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_zero_conditional.R
\name{pad_zero_conditional}
\alias{pad_zero_conditional}
\title{pad strings of length X with leading zeroes or any other char}
\usage{
pad_zero_conditional(str, str_len = 7, char_pad = "0")
}
\arguments{
\item{str}{string}

\item{str_len}{nchar of string for which 0s or other char will be added}

\item{char_pad}{char to be added, default 0}
}
\value{
returns a string vector
}
\description{
pad strings of length X with leading zeroes or any other char
}
