% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_png.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{PDF to PNG}
\usage{
pdf_to_png(pdf_file_path, output_dir, dpi = 600)
}
\arguments{
\item{pdf_file_path}{PDF file path}

\item{output_dir}{Directory where PNG should go}

\item{dpi}{DPI}
}
\description{
PDF to PNG
}
