% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_matchcolumns.R
\name{rbind.match.columns}
\alias{rbind.match.columns}
\title{Combining dataframes when the columns don’t match}
\usage{
\method{rbind}{match.columns}(input1, input2)
}
\arguments{
\item{input1}{first data frame}

\item{input2}{2nd data frame}
}
\description{
Combining dataframes when the columns don’t match
}
\examples{
rbind.match.columns(database.one, database.two)
https://amywhiteheadresearch.wordpress.com/2013/05/13/combining-dataframes-when-the-columns-dont-match/
}
