% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read all worksheets}
\usage{
read_excel_allsheets(filename, tibble = FALSE, ...)
}
\arguments{
\item{...}{arguments passed to read_excel}
}
\description{
Read all worksheets
}
