% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_somesheets.R
\name{read_excel_somesheets}
\alias{read_excel_somesheets}
\title{Read all worksheets}
\usage{
read_excel_somesheets(filename, tibble = FALSE, string_find = "", ...)
}
\arguments{
\item{string_find}{this is the string that is used for filtering sheet names}

\item{...}{arguments passed to read_excel}
}
\description{
Read all worksheets
}
