% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ags_old_to_new.R
\docType{data}
\name{reformen}
\alias{reformen}
\title{function to convert ags from old to new}
\format{
An object of class \code{data.frame} with 5457 rows and 4 columns.
}
\usage{
reformen
}
\arguments{
\item{ags_old_input}{OLD AGS back to 2006}
}
\value{
2022 AGS
}
\description{
function to convert ags from old to new
}
\keyword{datasets}
