% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_intersect.R
\name{rel_intersect}
\alias{rel_intersect}
\title{Checks the relative size of the intersection of two sets}
\usage{
rel_intersect(data1, data2, varname1, varname2)
}
\arguments{
\item{data1}{First data set}

\item{data2}{Second data set}

\item{varname1}{Name of variable in first data set}

\item{varname2}{Name of variable in second data set}
}
\value{
This returns some overlap statistics
}
\description{
Checks the relative size of the intersection of two sets
}
