% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_old_plots.R
\name{rremove_unused_plots}
\alias{rremove_unused_plots}
\title{Remove unused plot files}
\usage{
rremove_unused_plots(code_dir, plot_dir)
}
\arguments{
\item{code_dir}{The directory containing the R code files.}

\item{plot_dir}{The directory containing the PDF plot files.}
}
\value{
A character vector of the plots that were deleted.
}
\description{
This function identifies and removes PDF plots in the specified plot directory that are not referenced by any R scripts in the specified code directory.
}
\examples{
remove_unused_plots(
    code_dir = "path/to/code/directory",
    plot_dir = "path/to/plot/directory"
)

}
