% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_id_to_names.R
\name{state_id_to_names}
\alias{state_id_to_names}
\title{Conversion between German state IDs (AGS) and English names}
\usage{
state_id_to_names(state_id, how = "full_name")
}
\arguments{
\item{state_id}{state id vector}

\item{how}{character, either "full_name" or "short_name"}
}
\description{
Conversion between German state IDs (AGS) and English names
}
