% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_filter.R
\name{str_filter}
\alias{str_filter}
\title{Takes string, looks for expr and then finds matches and returns them}
\usage{
str_filter(string, pattern)
}
\arguments{
\item{string}{list of felm models to be passed to stargazer}

\item{pattern}{name of explanatory variable to be replaced}
}
\value{
Returns string
}
\description{
Takes string, looks for expr and then finds matches and returns them
}
