% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substr_right.R
\name{substrRight}
\alias{substrRight}
\title{Extracting the last n characters from a string in R}
\usage{
substrRight(x, n)
}
\description{
Extracting the last n characters from a string in R
}
\examples{
substrRight(x, 6)
"string"
}
