% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_vector.R
\name{summarize_vec}
\alias{summarize_vec}
\title{Summarize a vector}
\usage{
summarize_vec(vector)
}
\arguments{
\item{vector}{numeric vector to summarize}
}
\description{
Summarize a vector
}
\examples{

}
