% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_felm.R
\name{tidy_felm}
\alias{tidy_felm}
\title{Felm output summarize}
\usage{
tidy_felm(model, add_glance = T)
}
\arguments{
\item{model}{felm model}

\item{add_glance}{if T, glance output will be add to each line (note that not all default glance output will be returned)}
}
\description{
Felm output summarize
Main diff. to normal tidy is that this also gets number of obs
}
