% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_felm_list.R
\name{tidy_felm_list}
\alias{tidy_felm_list}
\title{Felm output summarize - list version}
\usage{
tidy_felm_list(model_list)
}
\arguments{
\item{model_list}{list of felm models}
}
\description{
Felm output summarize - list version
}
