% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_feols.R
\name{tidy_feols}
\alias{tidy_feols}
\title{feols output summarize for multiple outcomes}
\usage{
tidy_feols(model, ...)
}
\arguments{
\item{model}{feols model with multiple outcomes}

\item{...}{arguments passed to tidy_feols_single (see documentation)}
}
\description{
feols output summarize for multiple outcomes (should also work for just one outcome)
}
