% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_RD.R, R/tidy_felm.R
\name{tidy_rd}
\alias{tidy_rd}
\title{Summarize the Output from rdrobust}
\usage{
tidy_rd(model)

tidy_rd(model)
}
\arguments{
\item{model}{felm model}

\item{se_nr}{specify the standard error: 1 = conventional, 2 = Bias-Corrected, 3 = Robust}
}
\description{
Summarize the Output from rdrobust

Summarize the output from felm
}
\examples{




}
