% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_party_convert.R
\name{german_party_convert}
\alias{german_party_convert}
\title{Convert german party names (de or en) to proper German party names}
\usage{
german_party_convert(
  party_vec,
  far_right_label = "Far-right parties (incl. AfD)"
)
}
\arguments{
\item{party_vec}{vector of party names}
}
\description{
Convert german party names (de or en) to proper German party names
}
