% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hanno.R
\name{theme_hanno}
\alias{theme_hanno}
\title{This is my ggplot theme}
\usage{
theme_hanno(fontsize = 15, font = "default")
}
\arguments{
\item{fontsize}{Selecte fontsize, default is 15}

\item{font}{Selected font. Either 'default' or 'frutiger'.}
}
\value{
This is a custom ggplot theme. You can change the font size and the font family.
}
\description{
This is my ggplot theme
}
\examples{
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
geom_point() +
theme_hanno(fontsize = 15, font = 'frutiger')


}
