% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_felm.R
\name{tidy_felm}
\alias{tidy_felm}
\title{Felm output summarize}
\usage{
tidy_felm(model)
}
\arguments{
\item{model}{felm model}
}
\description{
Felm output summarize
Main diff. to normal tidy is that this also gets number of obs
}
