% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_RD.R
\name{tidy_rd}
\alias{tidy_rd}
\title{Summarize the output from rdrobust}
\usage{
tidy_rd(model, se_nr)
}
\arguments{
\item{se_nr}{specify the standard error: 1 = conventional, 2 = Bias-Corrected, 3 = Robust}
}
\description{
Summarize the Output from rdrobust
(works similar to tidy() for lm objects)
}
